+++ server/datastore/mysql/mysql.go
@@ -103,40 +103,42 @@ func (d *Datastore) MigrateData() error {
 	return nil
 }
 
-func (d *Datastore) MigrationStatus() error {
+func (d *Datastore) MigrationStatus() (kolide.MigrationStatus, error) {
 	if tables.MigrationClient.Migrations == nil || data.MigrationClient.Migrations == nil {
-		return errors.New("unexpected nil migrations list")
+		return 0, errors.New("unexpected nil migrations list")
 	}
 
 	lastTablesMigration, err := tables.MigrationClient.Migrations.Last()
 	if err != nil {
-		return errors.New("missing tables migrations")
+		return 0, errors.New("missing tables migrations")
 	}
 
 	currentTablesVersion, err := tables.MigrationClient.GetDBVersion(d.db.DB)
 	if err != nil {
-		return errors.New("cannot get table migration status")
-	}
-
-	if currentTablesVersion != lastTablesMigration.Version {
-		return errors.New("table migrations must be run")
+		return 0, errors.New("cannot get table migration status")
 	}
 
 	lastDataMigration, err := data.MigrationClient.Migrations.Last()
 	if err != nil {
-		return errors.New("missing data migrations")
+		return 0, errors.New("missing data migrations")
 	}
 
 	currentDataVersion, err := data.MigrationClient.GetDBVersion(d.db.DB)
 	if err != nil {
-		return errors.New("cannot get table migration status")
+		return 0, errors.New("cannot get table migration status")
 	}
 
-	if currentDataVersion != lastDataMigration.Version {
-		return errors.New("data migrations must be run")
-	}
+	switch {
+	case currentDataVersion == 0 && currentTablesVersion == 0:
+		return kolide.NoMigrationsCompleted, nil
 
-	return nil
+	case currentTablesVersion != lastTablesMigration.Version ||
+		currentDataVersion != lastDataMigration.Version:
+		return kolide.SomeMigrationsCompleted, nil
+
+	default:
+		return kolide.AllMigrationsCompleted, nil
+	}
 }
 
 // Drop removes database