+++ server/mock/datastore.go
@@ -38,8 +38,8 @@ func (m *Store) MigrateTables() error {
 func (m *Store) MigrateData() error {
 	return nil
 }
-func (m *Store) MigrationStatus() error {
-	return nil
+func (m *Store) MigrationStatus() (kolide.MigrationStatus, error) {
+	return 0, nil
 }
 func (m *Store) Name() string {
 	return "mock"